+++ back/src/routes/post.js
@@ -1,9 +1,9 @@
 const { Router } = require('express');
 const router = Router();
-const { Post } = require('../db');
+const { Post, UserInfo } = require('../db');
+// const UserInfo = require('../models/UserInfo');
 
-
-          // || /POST || //
+// || /POST || //
 router.get('', async (req, res) => {
   try {
     const allPost = await Post.findAll();
@@ -13,17 +13,34 @@ router.get('', async (req, res) => {
   }
 });
 
+// || POST/:ID || //
+router.get('/:id', async (req, res) => {
+  const id = req.params.id;
+  const post = await Post.findByPk(id);
+  const idUser = post.UserInfoId;
+  const postWithUser = await Post.findByPk(id, {
+    include: {
+      model: UserInfo,
+      attributes: ['name'],
+      where: {
+        id: idUser,
+      },
+    },
+  });
+  res.json(postWithUser);
+});
 
-          // || POST /POST || //
+// || POST /POST || //
 router.post('', async (req, res) => {
   try {
-    const { likes, powersGained, multimedia, description} = req.body;
-    if ((likes && powersGained, multimedia, description)) {
+    const { likes, powersGained, multimedia, description, UserInfoId } = req.body;
+    if (likes && powersGained && multimedia && description && UserInfoId) {
       const newPost = await Post.create({
         likes,
         powersGained,
         multimedia,
         description,
+        UserInfoId,
       });
       res.json(newPost);
     } else {
@@ -32,28 +49,30 @@ router.post('', async (req, res) => {
   } catch (error) {
     res.status(500).json({ message: error.message });
   }
-          // | AGREGAR RELACION CON USERID | //
 });
 
-
-          // || DELETE /POST || //
+// || DELETE /POST || //
 router.delete('/:id', async (req, res) => {
   try {
     const { id } = req.params;
-    await Post.destroy({
-      where: {
-        id,
-      },
-    });
-    res.sendStatus(204);
+    const post = await Post.findByPk(id);
+    if (post !== null) {
+      await Post.destroy({
+        where: {
+          id,
+        },
+      });
+      res.sendStatus(204);
+    } else {
+      throw new Error('the id does not exist');
+    }
   } catch (error) {
     return res.status(500).json({ message: error.message });
   }
 });
 
-
-          // || PUT /POST || //
-          // || MODIFICA TODO EL POST O SOLO ALGO PERO HAY QUE PONER TODO || //
+// || PUT /POST || //
+// || MODIFICA TODO EL POST O SOLO ALGO PERO HAY QUE PONER TODO || //
 // router.put('/:id', async (req, res) => {
 //   try {
 //     const { id } = req.params;
@@ -72,7 +91,7 @@ router.delete('/:id', async (req, res) => {
 //   }
 // });
 
-          // || MOFICIA SOLO LOS LIKES Y LOS POWERS || //
+// || MOFICIA SOLO LOS LIKES Y LOS POWERS || //
 router.put('/:id', async (req, res) => {
   try {
     const { id } = req.params;
